move `t_symbol_stack` into `t_symbol`
