+++ src/parse/t_symbol/t_symbol_stack/t_symbol_stack.c
